+++ src/main/java/com/enonic/harvest/harvestclient/models/User.java
@@ -6,10 +6,6 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 import javax.xml.bind.JAXBContext;
 import javax.xml.bind.Unmarshaller;
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlRootElement;
 import java.io.InputStream;
 import java.math.BigDecimal;
 import java.util.Date;
@@ -19,7 +15,7 @@ public class User
 {
 
     @JsonProperty("id")
-    private Integer id;
+    private Long id;
 
     @JsonProperty("email")
     private String email;
@@ -61,12 +57,12 @@ public class User
     private Date createdAt;
 
 
-    public Integer getId()
+    public Long getId()
     {
         return id;
     }
 
-    public void setId(Integer id)
+    public void setId(Long id)
     {
         this.id = id;
     }
